requirements.txt
asgiref==3.8.1
babel==2.17.0
datetime-truncate==1.1.1
diff-match-patch==20241021
Django==5.1.7
django-bower==5.2.0
django-colorfield==0.12.0
django-environ==0.12.0
django-import-export==4.3.7
django-js-asset==3.1.2
django-memoize==2.3.1
django-money==3.5.3
django-mptt==0.16.0
django-mptt-admin==2.7.2
django-multiselectfield==0.1.13
django-nvd3==0.10.1
django-treenode==0.22.2
django-unfold==0.51.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==11.1.0
py-moneyed==3.0
python-dateutil==2.9.0.post0
python-nvd3==0.16.0
python-slugify==8.0.4
setuptools==76.0.0
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
text-unidecode==1.3
typing_extensions==4.12.2
